chore: Reorganise search file structure (#3563)
